requirements.txt
Flask==2.0.2
openai==0.27.0
python-dotenv==1.0.0
Werkzeug==2.2.2